refactor: extract popover overlay fix to the composible [WTEL-9748](https://webitel.atlassian.net/browse/WTEL-9748)
